event_proc/requirements.txt
pymongo
pulsar-client
pandas
numpy
matplotlib
geopy
shapely